+++ ports/atmel-samd/clocks.c
@@ -56,77 +56,6 @@ uint8_t find_free_gclk(uint16_t divisor) {
     return 0xff;
 }
 
-bool gclk_enabled(uint8_t gclk) {
-    #ifdef SAMD51
-    return GCLK->GENCTRL[gclk].bit.GENEN;
-    #endif
-    #ifdef SAMD21
-    common_hal_mcu_disable_interrupts();
-    // Explicitly do a byte write so the peripheral knows we're just wanting to read the channel
-    // rather than write to it.
-    *((uint8_t*) &GCLK->GENCTRL.reg) = gclk;
-    while (GCLK->STATUS.bit.SYNCBUSY == 1) {}
-    bool enabled = GCLK->GENCTRL.bit.GENEN;
-    common_hal_mcu_enable_interrupts();
-    return enabled;
-    #endif
-}
-
-void disable_gclk(uint8_t gclk) {
-    #ifdef SAMD51
-    while ((GCLK->SYNCBUSY.vec.GENCTRL & (1 << gclk)) != 0) {}
-    GCLK->GENCTRL[gclk].bit.GENEN = false;
-    while ((GCLK->SYNCBUSY.vec.GENCTRL & (1 << gclk)) != 0) {}
-    #endif
-    #ifdef SAMD21
-    while (GCLK->STATUS.bit.SYNCBUSY == 1) {}
-    GCLK->GENCTRL.reg = GCLK_GENCTRL_ID(gclk);
-    while (GCLK->STATUS.bit.SYNCBUSY == 1) {}
-    #endif
-}
-
-void connect_gclk_to_peripheral(uint8_t gclk, uint8_t peripheral) {
-    #ifdef SAMD21
-    GCLK->CLKCTRL.reg = GCLK_CLKCTRL_ID(peripheral) | GCLK_CLKCTRL_GEN(gclk) | GCLK_CLKCTRL_CLKEN;
-    #endif
-    #ifdef SAMD51
-    GCLK->PCHCTRL[peripheral].reg = GCLK_PCHCTRL_CHEN | GCLK_PCHCTRL_GEN(gclk);
-    while(GCLK->SYNCBUSY.reg != 0) {}
-    #endif
-}
-
-void disconnect_gclk_from_peripheral(uint8_t gclk, uint8_t peripheral) {
-    #ifdef SAMD21
-    GCLK->CLKCTRL.reg = GCLK_CLKCTRL_ID(peripheral) | GCLK_CLKCTRL_GEN(gclk);
-    #endif
-    #ifdef SAMD51
-    GCLK->PCHCTRL[peripheral].reg = 0;
-    #endif
-}
-
-void enable_clock_generator(uint8_t gclk, uint8_t source, uint16_t divisor) {
-    #ifdef SAMD21
-    GCLK->GENDIV.reg = GCLK_GENDIV_ID(gclk) | GCLK_GENDIV_DIV(divisor);
-    GCLK->GENCTRL.reg = GCLK_GENCTRL_ID(gclk) | GCLK_GENCTRL_SRC(source) | GCLK_GENCTRL_GENEN;
-    while (GCLK->STATUS.bit.SYNCBUSY != 0) {}
-    #endif
-    #ifdef SAMD51
-    GCLK->GENCTRL[gclk].reg = GCLK_GENCTRL_SRC(source) | GCLK_GENCTRL_DIV(divisor) | GCLK_GENCTRL_GENEN;
-    while ((GCLK->SYNCBUSY.vec.GENCTRL & (1 << gclk)) != 0) {}
-    #endif
-}
-
-void disable_clock_generator(uint8_t gclk) {
-    #ifdef SAMD21
-    GCLK->GENCTRL.reg = GCLK_GENCTRL_ID(gclk);
-    while (GCLK->STATUS.bit.SYNCBUSY != 0) {}
-    #endif
-    #ifdef SAMD51
-    GCLK->GENCTRL[gclk].reg = 0;
-    while ((GCLK->SYNCBUSY.vec.GENCTRL & (1 << gclk)) != 0) {}
-    #endif
-}
-
 void reset_gclks(void) {
     // Never reset GCLK0 because its used for the core
     #if CONF_GCLK_GEN_1_GENEN == 0
